chore: custom-errors as its own example
